% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mObject.R
\name{create a square}
\alias{create a square}
\alias{square}
\title{A square.}
\usage{
square(side_length = 2, ...)
}
\arguments{
\item{side_length}{The length of the side of the square. Default is set to \code{2}.}

\item{...}{Additional arguments to be passed to \code{\link{Arc}} object.}
}
\description{
A square.
}
\examples{
scene <- init_scene()

anim <- square() |>
             flip(LEFT()) |>
             rotate(-3 * pi / 8) |>
             create()

play(scene, anim)

scene |> render()

}
