% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mObject.R
\name{create a triangle}
\alias{create a triangle}
\alias{triangle}
\title{An equilateral triangle}
\usage{
triangle(...)
}
\arguments{
\item{...}{Additional arguments to be passed to \code{\link{RegularPolygon}} object.}
}
\description{
An equilateral triangle
}
\examples{
scene <- init_scene()

anim <- triangle() |>
             flip(LEFT()) |>
             rotate(-3 * pi / 8) |>
             create()

play(scene, anim)

scene |> render()

}
