% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animations.R
\name{create mObject}
\alias{create mObject}
\alias{create}
\title{Incrementally show a mObject}
\usage{
create(mobject)
}
\arguments{
\item{mobject}{The \code{\link{mObject}} to animate.}
}
\description{
Incrementally show a mObject
}
\examples{
scene <- init_scene()

anim <- circle() |>
             flip(LEFT()) |>
             rotate(-3 * pi / 8) |>
             create()

play(scene, anim)

scene |> render()

}
