% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animations.R
\name{display mObject}
\alias{display mObject}
\alias{display}
\title{Incrementally display a mObject}
\usage{
display(mobject)
}
\arguments{
\item{mobject}{The \code{\link{mObject}} to animate.}
}
\description{
Incrementally display a mObject
}
\examples{
scene <- init_scene()

anim <- circle() |>
             flip(LEFT()) |>
             rotate(-3 * pi / 8) |>
             display()

play(scene, anim)

scene |> render()

}
