% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animations.R
\name{fade out mObject}
\alias{fade out mObject}
\alias{fadeout}
\title{Fade-Out effect for mObjects}
\usage{
fadeout(mobject)
}
\arguments{
\item{mobject}{The \code{\link{mObject}} to animate.}
}
\description{
Fade-Out effect for mObjects
}
\examples{
scene <- init_scene()

anim <- fadeout(circle(), shift= DOWN)

play(scene, anim)

scene |> render()

}
