% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendering.R
\name{play}
\alias{play}
\title{Play an animation}
\usage{
play(scene, animation)
}
\arguments{
\item{scene}{A \code{\link{Scene}} object.}

\item{animation}{A \code{\link{Animation}} object.}
}
\description{
Play an animation
}
