% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translations.R
\name{rotate a mObject}
\alias{rotate a mObject}
\alias{rotate}
\title{Rotate a mObject by a vector}
\usage{
rotate(mobject, angle)
}
\arguments{
\item{mobject}{The \code{\link{mObject}} to animate.}

\item{angle}{The angle by which the mobject is rotated.}
}
\description{
Rotate a mObject by a vector
}
\examples{
scene <- init_scene()

anim <- circle() |>
           flip(LEFT()) |>
           rotate(-3 * pi / 8) |>
           create()

play(scene, anim)

scene |> render()

}
