% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translations.R
\name{translate a mObject}
\alias{translate a mObject}
\alias{flip}
\title{Translate a mObject by a vector}
\usage{
flip(mobject, direction = UP)
}
\arguments{
\item{mobject}{The \code{\link{mObject}} to animate.}

\item{direction}{The vector by which the mobject is translated. Default is set to \code{\link{UP}}.}
}
\description{
Translate a mObject by a vector
}
\examples{
scene <- init_scene()

anim <- circle() |>
            flip(LEFT()) |>
            translate(UP() * 2) |>
            create()

play(scene, anim)

scene |> render()

}
