% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animations.R
\name{transmute mObject}
\alias{transmute mObject}
\alias{transmute}
\title{Transmutes a mObject into a target mObject}
\usage{
transmute(mobject, target)
}
\arguments{
\item{mobject}{The \code{\link{mObject}} to animate.}

\item{target_object}{The target of the transformation.}

\item{path_func}{A function defining the path that the points of the \code{\link{mObject}} are being moved long until they match the points of the \code{\link{target_mobject}}.}

\item{path}{path_arc_centers The center of the circular arcs along which the points of \code{\link{mObject}} are moved by the transformation. If this is set and \code{\link{path_func}} is not set, then a \code{\link{path_along_circles}} path will be generated
using the \code{\link{path_arc}}parameters and stored in \code{\link{path_func}} If \code{\link{path_func}}i s set, this and the
other \code{\link{path_arc}}fields are set as attributes, but a \code{\link{path_func}} is not generated from it. Controls which mobject is replaced when the transformation is complete. If set to True, \code{\link{mObject}} will be removed from the scene and \code{\link{target_mobject}} will replace it. Otherwise, \code{\link{target_mobject}} is never added and \code{\link{mObject}} just takes its shape.}
}
\description{
Transmutes a mObject into a target mObject
}
\examples{
scene <- init_scene()

pink <- "#D147BD"

anim = square() |>
     transmute( circle() |>
     set_fill(pink, opacity=0.5) )

play(scene, anim)

scene |> render()

}
