% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{transmutes mObjects}
\alias{transmutes mObjects}
\alias{set_fill}
\title{Set the fill color and fill opacity of a mObject.}
\usage{
set_fill(mobject, color = NULL, opacity = 1, family = TRUE)
}
\arguments{
\item{color}{Fill color of the \code{\link{mObject}}.}

\item{opacity}{Fill opacity of the \code{\link{mObject}}.}

\item{family}{If \code{True}, the fill color of all submobjects is also set. Default set as TRUE.}
}
\description{
Set the fill color and fill opacity of a mObject.
}
\examples{
scene <- init_scene()

pink <- "#D147BD"

anim = square() |>
     transmute( circle() |>
     set_fill(color=pink, opacity=0.5) )

play(scene, anim)

scene |> render()

}
